+++ attributes/default.rb
@@ -65,6 +65,9 @@ default['ntp']['tinker']['panic'] = 1000
 default['ntp']['tinker']['step'] = 0.128
 default['ntp']['tinker']['stepout'] = 900
 
+# Set to true if using ntp < 4.2.8 or any unpatched ntp version to mitigate CVE-2014-9293 / CVE-2014-9294 / CVE-2014-9295
+default['ntp']['localhost']['noquery'] = false
+
 # overrides on a platform-by-platform basis
 case node['platform_family']
 when 'debian'