+++ CHANGELOG.md
@@ -7,6 +7,8 @@ v1.7.2 (2014-12-10)
 - Chefspec 4.0 updates
 - Rubocop updates
 - **[PR #85](https://github.com/gmiranda23/ntp/pull/85)** - Update leapseconds for June 2015 leapsecond
+- **[PR #70](https://github.com/gmiranda23/ntp/pull/70)** - Allow setting tinker options in attributes
+- **[PR #84](https://github.com/gmiranda23/ntp/pull/84)** - Add attributes for tinker option customization
 
 
 v1.7.0 (2014-12-10)